requirements.txt
streamlit>=1.30.0
supabase>=2.0.0
python-dotenv>=1.0.0
pyproj>=3.6.0
pandas>=2.0.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.18.0
qrcode[pil]>=7.4
reportlab>=4.0
openpyxl>=3.1.0
python-docx>=1.0.0
gotrue>=2.0.0
